rename modules according to weka lib, proposal for #8
