+++ features/step_definitions/weka_pipeline.rb
@@ -12,7 +12,7 @@ Then(/^I want to parse them$/) do
 end
 
 Then(/^I want to filter them using the unsupervised filter "(.*?)"$/) do |arg1|
-  filter = Weka::Filter::Unsupervised::Attribute::StringToWordVector.new
+  filter = Weka::Filters::Unsupervised::Attribute::StringToWordVector.new
   filter.set do
     data $training_set
   end
@@ -23,7 +23,7 @@ Then(/^I want to filter them using the unsupervised filter "(.*?)"$/) do |arg1|
 end
 
 Then(/^I want to build a "(.*?)" classifier using training set data$/) do |arg1|
-  $bayes = Weka::Classifier::Bayes::ComplementNaiveBayes::Base.new do
+  $bayes = Weka::Classifiers::Bayes::ComplementNaiveBayes::Base.new do
     set_data $filtered_training_set
     set_class_index 0
   end
@@ -31,7 +31,7 @@ Then(/^I want to build a "(.*?)" classifier using training set data$/) do |arg1|
 end
 
 Then(/^I want to evaluate the performance of the classifier on the test set$/) do
-  @evaluator = Weka::Classifier::Evaluation.new $filtered_training_set
+  @evaluator = Weka::Classifiers::Evaluation.new $filtered_training_set
   @evaluator.evaluate_model($bayes,$filtered_test_set)
 end
