+++ lib/ruby-band/weka/classifiers/functions/functions.rb
@@ -1,15 +1,11 @@
-$:.unshift File.dirname(__FILE__)
+require 'ruby-band/weka/classifiers/class_builder'
 
-require_relative '../class_builder'
-
-#
-# The Weka::Classifier::Functions module contains classifiers from the 'weka.classifiers.functions' package
+# The Weka::Classifiers::Functions module contains classifiers from the 'weka.classifiers.functions' package
 # New weka classes can be defined by adding them to the build_classes method.
-#
 module Weka
-	module Classifier
-		module Functions
-      include ClassBuilder::Classifiers
+  module Classifiers
+    module Functions
+      include ClassBuilder
 
       build_classes :LinearRegression,
                     :PLSClassifier,
