+++ lib/ruby-band/weka/classifiers/mi/mi.rb
@@ -1,15 +1,12 @@
-$:.unshift File.dirname(__FILE__)
+require 'ruby-band/weka/classifiers/class_builder'
 
-require_relative '../class_builder'
-
-#
-# The Weka::Classifier::Mi module contains classifiers from the 'weka.classifiers.mi' package
+# The Weka::Classifiers::Mi module contains classifiers from the 'weka.classifiers.mi' package
 # New weka classes can be defined by adding them to the build_classes method.
 #
 module Weka
-	module Classifier
-		module Mi
-      include ClassBuilder::Classifiers
+  module Classifiers
+    module Mi
+      include ClassBuilder
 
       build_classes :CitationKNN,
                     :MDD,
@@ -23,6 +20,6 @@ module Weka
                     :MISVM,
                     :MIWrapper,
                     :SimpleMI
-		end
+    end
   end
 end
