+++ lib/ruby-band/weka/filters/unsupervised/attribute/attribute.rb
@@ -1,12 +1,9 @@
-$:.unshift(File.expand_path(File.join(File.dirname(__FILE__),"../")))
-require 'java'
-require 'unsupervised_utils'
+require 'ruby-band/weka/filters/unsupervised/utils'
 
 module Weka
-  module Filter
+  module Filters
   	module Unsupervised
   		module Attribute
-
   			java_import "weka.filters.unsupervised.attribute.Add"
         java_import "weka.filters.unsupervised.attribute.AddCluster"
   			java_import "weka.filters.unsupervised.attribute.Center"
@@ -18,72 +15,74 @@ module Weka
   			java_import "weka.filters.unsupervised.attribute.Standardize"
   			java_import "weka.filters.unsupervised.attribute.PrincipalComponents"
         java_import "weka.filters.unsupervised.attribute.StringToWordVector"
-  			java_import "weka.core.Utils"
-  			java_import "weka.filters.Filter"
 
   			class Add
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
         class AddCluster
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
+
           alias_method :clusterer, :set_clusterer
+
           def set_clusterer(index)
             set_clusterer(index)
           end
         end
 
   			class Center
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
         class ClusterMembership
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
         end
 
   			class Discretize
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
   			class NominalToString
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
   			class Normalize
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
   			class PrincipalComponents
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
       
   			class Remove
+          include Weka::Filters::Unsupervised::Utils
+
           alias_method :attribute_indices, :setAttributeIndices
+
           def setAttributeIndices(index)
             setAttributeIndices(index)
           end
-          include Unsupervised_Util
   			end
 
   			class Standardize
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
         class StringToWordVector
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
         end
 
-        Weka::Filter::Unsupervised::Attribute::PrincipalComponents.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::Add.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::AddCluster.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::ClusterMembership.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::Standardize.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::Remove.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::Normalize.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::Discretize.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::Center.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::NominalToString.__persistent__ = true
-        Weka::Filter::Unsupervised::Attribute::StringToWordVector.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::PrincipalComponents.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::Add.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::AddCluster.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::ClusterMembership.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::Standardize.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::Remove.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::Normalize.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::Discretize.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::Center.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::NominalToString.__persistent__ = true
+        Weka::Filters::Unsupervised::Attribute::StringToWordVector.__persistent__ = true
   		end
   	end
   end
