+++ lib/ruby-band/weka/filters/unsupervised/instance/instance.rb
@@ -1,13 +1,9 @@
-$:.unshift(File.expand_path(File.join(File.dirname(__FILE__),"../")))
-require 'java'
-require 'unsupervised_utils'
+require 'ruby-band/weka/filters/unsupervised/utils'
 
 module Weka
-  module Filter
+  module Filters
   	module Unsupervised
   		module Instance
-  			java_import "weka.core.Utils"
-  			java_import "weka.filters.Filter"
   			java_import "weka.filters.unsupervised.instance.Normalize"
   			java_import "weka.filters.unsupervised.instance.NonSparseToSparse"
   			java_import "weka.filters.unsupervised.instance.SparseToNonSparse"
@@ -16,29 +12,34 @@ module Weka
         java_import "weka.filters.unsupervised.instance.RemoveWithValues"
 
   			class Normalize
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
   			class NonSparseToSparse
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
   			class RemoveRange
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
         class RemoveWithValues
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
         end
 
   			class SparseToNonSparse
-          include Unsupervised_Util
+          include Weka::Filters::Unsupervised::Utils
   			end
 
   			class SubsetByExpression
-          include Unsupervised_Util
-  				def description
-            begin; puts globalInfo.split("\n").each {|line| puts line}; rescue; NoMethodError; puts "Sorry, no description available for this filter"; end
+          include Weka::Filters::Unsupervised::Utils
+
+          def description
+            begin
+              puts globalInfo.split("\n").each {|line| puts line}
+            rescue NoMethodError
+              puts "Sorry, no description available for this filter"
+            end
   				end
   			end
