+++ ruby-band.gemspec
@@ -2,14 +2,17 @@
 # DO NOT EDIT THIS FILE DIRECTLY
 # Instead, edit Jeweler::Tasks in Rakefile, and run 'rake gemspec'
 # -*- encoding: utf-8 -*-
+# stub: ruby-band 0.1.13 ruby lib
+# stub: ext/mkrf_conf.rb
 
 Gem::Specification.new do |s|
   s.name = "ruby-band"
   s.version = "0.1.13"
 
   s.required_rubygems_version = Gem::Requirement.new(">= 0") if s.respond_to? :required_rubygems_version=
+  s.require_paths = ["lib"]
   s.authors = ["arrigonialberto86"]
-  s.date = "2014-01-21"
+  s.date = "2014-02-01"
   s.description = "Data mining and machine learning algorithms for JRuby "
   s.email = "arrigonialberto86@gmail.com"
   s.executables = ["ruby-band"]
@@ -67,33 +70,29 @@ Gem::Specification.new do |s|
     "lib/ruby-band/core/type/instances.rb",
     "lib/ruby-band/core/type/utils.rb",
     "lib/ruby-band/weka.rb",
-    "lib/ruby-band/weka/attribute_selection/attribute_selection_utils.rb",
     "lib/ruby-band/weka/attribute_selection/evaluators.rb",
     "lib/ruby-band/weka/attribute_selection/search.rb",
+    "lib/ruby-band/weka/attribute_selection/utils.rb",
     "lib/ruby-band/weka/classifiers/bayes/bayes.rb",
-    "lib/ruby-band/weka/classifiers/bayes/bayes_utils.rb",
+    "lib/ruby-band/weka/classifiers/class_builder.rb",
     "lib/ruby-band/weka/classifiers/evaluation.rb",
     "lib/ruby-band/weka/classifiers/functions/functions.rb",
-    "lib/ruby-band/weka/classifiers/functions/functions_utils.rb",
     "lib/ruby-band/weka/classifiers/lazy/lazy.rb",
-    "lib/ruby-band/weka/classifiers/lazy/lazy_utils.rb",
     "lib/ruby-band/weka/classifiers/mi/mi.rb",
-    "lib/ruby-band/weka/classifiers/mi/mi_utils.rb",
     "lib/ruby-band/weka/classifiers/rules/rules.rb",
-    "lib/ruby-band/weka/classifiers/rules/rules_utils.rb",
     "lib/ruby-band/weka/classifiers/trees/trees.rb",
-    "lib/ruby-band/weka/classifiers/trees/trees_utils.rb",
+    "lib/ruby-band/weka/classifiers/utils.rb",
     "lib/ruby-band/weka/clusterers/clusterers.rb",
-    "lib/ruby-band/weka/clusterers/clusterers_utils.rb",
+    "lib/ruby-band/weka/clusterers/utils.rb",
     "lib/ruby-band/weka/db/DatabaseUtils_mysql",
     "lib/ruby-band/weka/db/DatabaseUtils_postgresql",
     "lib/ruby-band/weka/db/db.rb",
     "lib/ruby-band/weka/filters/supervised/attribute/attribute.rb",
     "lib/ruby-band/weka/filters/supervised/instance/instance.rb",
-    "lib/ruby-band/weka/filters/supervised/supervised_utils.rb",
+    "lib/ruby-band/weka/filters/supervised/utils.rb",
     "lib/ruby-band/weka/filters/unsupervised/attribute/attribute.rb",
     "lib/ruby-band/weka/filters/unsupervised/instance/instance.rb",
-    "lib/ruby-band/weka/filters/unsupervised/unsupervised_utils.rb",
+    "lib/ruby-band/weka/filters/unsupervised/utils.rb",
     "resources/ReutersGrain-test.arff",
     "resources/ReutersGrain-train.arff",
     "resources/weather.csv",
@@ -108,12 +107,11 @@ Gem::Specification.new do |s|
   ]
   s.homepage = "http://github.com/arrigonialberto86/ruby-band"
   s.licenses = ["MIT"]
-  s.require_paths = ["lib"]
-  s.rubygems_version = "1.8.24"
+  s.rubygems_version = "2.2.1"
   s.summary = "Data mining algorithms for JRuby"
 
   if s.respond_to? :specification_version then
-    s.specification_version = 3
+    s.specification_version = 4
 
     if Gem::Version.new(Gem::VERSION) >= Gem::Version.new('1.2.0') then
       s.add_runtime_dependency(%q<atomic>, [">= 0"])
